refactor(suite-common): split Tron stake thunks into per-action folders
