+++ suite-common/wallet-core/src/index.ts
@@ -70,7 +70,12 @@ export * from './stablecoin-yield/stablecoinYieldTypes';
 export * from './stablecoin-yield/stablecoinYieldUtils';
 export * from './stake/tron/tronStakeReducer';
 export * from './stake/tron/tronStakeSelectors';
-export * from './stake/tron/tronStakeThunks';
+export { composeTronFreezeFeeLevelsThunk } from './stake/tron/actions/freeze/composeFreeze';
+export { submitTronFreezeThunk } from './stake/tron/actions/freeze/submitFreeze';
+export { composeTronVoteFeeLevelsThunk } from './stake/tron/actions/vote/composeVote';
+export { submitTronVoteThunk } from './stake/tron/actions/vote/submitVote';
+export { composeTronUnstakeFeeLevelsThunk } from './stake/tron/actions/unstake/composeUnstake';
+export { submitTronUnstakeThunk } from './stake/tron/actions/unstake/submitUnstake';
 export * from './stake/tron/tronStakeTypes';
 export * from './stake/tron/tronStakeUtils';
 export * from './token/stellarTokenThunks';
