+++ suite-common/wallet-core/src/stake/tron/shared/constants.ts
@@ -1,0 +1,4 @@
+export const TRON_STAKE_MODULE = '@common/wallet-core/tron-stake';
+
+export const TRON_DUMMY_BLOCK_HASH = '0'.repeat(64);
+export const TRON_DUMMY_BLOCK_HEIGHT = 0;
